requirements.txt
fastapi==0.73.0
websockets==10.1
websocket-client==1.2.3
unicorn==1.0.3
python-multipart==0.0.5
pandas==1.4.0
numpy==1.22.1
yfinance==0.1.70
virtualenv==20.13.0
SQLALchemy==1.4.31
pip==22.0.3
alpaca-trade-api==1.5.0
# ta-lib==0.4.24
requests==2.27.1
# pysqlite==2.8.3




